feat(breaking)!: invert the logic for breaking change
with this change, if the user answers yes to the problem, the change would not
be marked as a breaking change. if, however, the user answers no to the
problem, then the change would be marked as a breaking change. this way, to
achieve the desired behavior, user would only need press yes all the way.